Update provisioner.sh
I dont know if I'm too new here, and I had to add `helm repo add isovalent https://helm.isovalent.com/` for my CUPE environment

I think it's ok to hardcode the name isovalent/cilium and URL here since it's that's all the possiblity that I can think of for now. we might want to set it as var in the future if we need to add some other repo for other purpose.